Move drivers back to their own (optionally dynamic)
This allows for reuse from projects dynamically linked to libeuicc. Note
that we don't reintroduce dlopen() based drivers here.

All backends except stdio have been made optional using CMake options.
C-side macros in driver.c have been adjusted to always mean enabling the
corresponding backend when defined.

Note that the GBinder backend does not need to distinguish between the
current HIDL version and a future AIDL implementation. Both will have
the same dependencies and will probably fall back on to each other
automatically.
