+++ CMakeLists.txt
@@ -44,4 +44,5 @@ endif()
 
 add_subdirectory(cjson)
 add_subdirectory(euicc)
+add_subdirectory(driver)
 add_subdirectory(src)
