+++ driver/CMakeLists.txt
@@ -1,0 +1,56 @@
+option(LPAC_DYNAMIC_DRIVERS "Build lpac/libeuicc driver backends as a dynamic library" OFF)
+
+option(LPAC_WITH_APDU_PCSC "Build APDU PCSC Backend (requires PCSC libraries)" ON)
+option(LPAC_WITH_APDU_AT "Build APDU AT Backend" ON)
+option(LPAC_WITH_APDU_GBINDER "Build APDU Gbinder backend for libhybris devices (requires gbinder headers)" OFF)
+
+option(LPAC_WITH_HTTP_CURL "Build HTTP Curl interface" ON)
+
+aux_source_directory(${CMAKE_CURRENT_SOURCE_DIR} DIR_INTERFACE_SRCS)
+if(LPAC_DYNAMIC_DRIVERS)
+    add_library(euicc-drivers SHARED ${DIR_INTERFACE_SRCS})
+else()
+    add_library(euicc-drivers STATIC ${DIR_INTERFACE_SRCS})
+endif()
+target_link_libraries(euicc-drivers euicc cjson-static)
+target_include_directories(euicc-drivers PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+
+target_sources(euicc-drivers PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/apdu/stdio.c)
+target_sources(euicc-drivers PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/http/stdio.c)
+
+if(LPAC_WITH_APDU_PCSC)
+    set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -DLPAC_WITH_APDU_PCSC")
+    target_sources(euicc-drivers PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/apdu/pcsc.c)
+    if(WIN32)
+        target_link_libraries(euicc-drivers winscard)
+    elseif(APPLE)
+        target_link_libraries(euicc-drivers "-framework PCSC")
+    else()
+        find_package(PCSCLite)
+        target_link_libraries(euicc-drivers PCSCLite::PCSCLite)
+    endif()
+endif()
+
+if(LPAC_WITH_APDU_AT)
+    set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -DLPAC_WITH_APDU_AT")
+    target_sources(euicc-drivers PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/apdu/at.c)
+endif()
+
+if(LPAC_WITH_APDU_GBINDER)
+    set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -DLPAC_WITH_APDU_GBINDER")
+    target_sources(euicc-drivers PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/apdu/gbinder_hidl.c)
+    find_package(PkgConfig REQUIRED)
+    pkg_check_modules(GBINDER REQUIRED IMPORTED_TARGET libgbinder)
+    pkg_check_modules(GLIB REQUIRED IMPORTED_TARGET glib-2.0)
+    target_link_libraries(euicc-drivers PkgConfig::GBINDER PkgConfig::GLIB)
+endif()
+
+if(LPAC_WITH_HTTP_CURL)
+    set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -DLPAC_WITH_HTTP_CURL")
+    target_sources(euicc-drivers PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/http/curl.c)
+    find_package(curl)
+    target_link_libraries(euicc-drivers curl)
+    if(WIN32)
+        target_link_libraries(euicc-drivers ${DL_LIBRARY})
+    endif()
+endif()
