+++ driver/driver.c
@@ -4,33 +4,33 @@
 #include <stdlib.h>
 #include <string.h>
 
-#ifdef LPAC_WITH_APDU_GBINDER_HIDL
+#ifdef LPAC_WITH_APDU_GBINDER
 #include "driver/apdu/gbinder_hidl.h"
 #endif
 
-#ifndef LPAC_WITHOUT_APDU_PCSC
+#ifdef LPAC_WITH_APDU_PCSC
 #include "driver/apdu/pcsc.h"
 #endif
-#ifndef LPAC_WITHOUT_APDU_AT
+#ifdef LPAC_WITH_APDU_AT
 #include "driver/apdu/at.h"
 #endif
-#ifndef LPAC_WITHOUT_HTTP_CURL
+#ifdef LPAC_WITH_HTTP_CURL
 #include "driver/http/curl.h"
 #endif
 #include "driver/apdu/stdio.h"
 #include "driver/http/stdio.h"
 
 static const struct lpac_driver *drivers[] = {
-#ifdef LPAC_WITH_APDU_GBINDER_HIDL
+#ifdef LPAC_WITH_APDU_GBINDER
     &driver_apdu_gbinder_hidl,
 #endif
-#ifndef LPAC_WITHOUT_APDU_PCSC
+#ifdef LPAC_WITH_APDU_PCSC
     &driver_apdu_pcsc,
 #endif
-#ifndef LPAC_WITHOUT_APDU_AT
+#ifdef LPAC_WITH_APDU_AT
     &driver_apdu_at,
 #endif
-#ifndef LPAC_WITHOUT_HTTP_CURL
+#ifdef LPAC_WITH_HTTP_CURL
     &driver_http_curl,
 #endif
     &driver_apdu_stdio,
@@ -38,19 +38,13 @@ static const struct lpac_driver *drivers[] = {
     NULL,
 };
 
-struct euicc_apdu_interface driver_interface_apdu;
-struct euicc_http_interface driver_interface_http;
-static struct applet_entry applet_apdu = {
-    .name = "apdu",
-    .main = NULL,
-};
-static struct applet_entry applet_http = {
-    .name = "http",
-    .main = NULL,
-};
-
 static const struct lpac_driver *_driver_apdu = NULL;
 static const struct lpac_driver *_driver_http = NULL;
+
+struct euicc_apdu_interface driver_interface_apdu;
+struct euicc_http_interface driver_interface_http;
+int (*driver_main_apdu)(int argc, char **argv) = NULL;
+int (*driver_main_http)(int argc, char **argv) = NULL;
 
 static const struct lpac_driver *_find_driver(enum lpac_driver_type type, const char *name)
 {
@@ -92,8 +86,8 @@ int driver_init()
     _driver_apdu->init(&driver_interface_apdu);
     _driver_http->init(&driver_interface_http);
 
-    applet_apdu.main = _driver_apdu->main;
-    applet_http.main = _driver_http->main;
+    driver_main_apdu = _driver_apdu->main;
+    driver_main_http = _driver_http->main;
 
     return 0;
 }
@@ -109,18 +103,3 @@ void driver_fini()
         _driver_http->fini();
     }
 }
-
-static int dlsym_interface_applet_main(int argc, char **argv)
-{
-    static const struct applet_entry *applets[] = {
-        &applet_apdu,
-        &applet_http,
-        NULL,
-    };
-    return applet_entry(argc, argv, applets);
-}
-
-struct applet_entry driver_applet = {
-    .name = "driver",
-    .main = dlsym_interface_applet_main,
-};
