+++ driver/driver.h
@@ -2,7 +2,6 @@
 #include <stddef.h>
 #include <inttypes.h>
 #include <euicc/interface.h>
-#include <applet.h>
 
 enum lpac_driver_type
 {
@@ -21,7 +20,8 @@ struct lpac_driver
 
 extern struct euicc_apdu_interface driver_interface_apdu;
 extern struct euicc_http_interface driver_interface_http;
-extern struct applet_entry driver_applet;
+extern int (*driver_main_apdu)(int argc, char **argv);
+extern int (*driver_main_http)(int argc, char **argv);
 
 int driver_init(void);
 void driver_fini(void);
