+++ src/CMakeLists.txt
@@ -11,19 +11,8 @@ aux_source_directory(${CMAKE_CURRENT_SOURCE_DIR}/applet/chip DIR_LPAC_SRCS)
 aux_source_directory(${CMAKE_CURRENT_SOURCE_DIR}/applet/notification DIR_LPAC_SRCS)
 aux_source_directory(${CMAKE_CURRENT_SOURCE_DIR}/applet/profile DIR_LPAC_SRCS)
 
-aux_source_directory(${CMAKE_CURRENT_SOURCE_DIR}/driver DIR_LPAC_SRCS)
-aux_source_directory(${CMAKE_CURRENT_SOURCE_DIR}/driver/apdu DIR_LPAC_SRCS)
-aux_source_directory(${CMAKE_CURRENT_SOURCE_DIR}/driver/http DIR_LPAC_SRCS)
-
-find_package(PCSCLite)
-find_package(curl)
-
 add_executable(lpac ${DIR_LPAC_SRCS})
-set_target_properties(lpac PROPERTIES
-    RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/output"
-    BUILD_RPATH "${RPATH_BINARY_PATH}"
-)
-target_link_libraries(lpac euicc PCSCLite::PCSCLite curl ${DL_LIBRARY})
+target_link_libraries(lpac euicc euicc-drivers)
 target_include_directories(lpac PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>)
 
 find_package(Git)
